Print bootkube journal if it fails
In previous attempt in e852117992a4a5ae13104a57ebb07dac71f9ed0f, I
missed the fact that Terraform remote-exec only prints stderr to the
user and stdout is ignored. This means, if we redirect journalctl output
to stderr, then it should be printed which will improve debugging
experience.

Refs #440

Signed-off-by: Mateusz Gozdek <mateusz@kinvolk.io>